7902389: Refactoring: misc part 2

+++ src/classes/com/sun/tdk/signaturetest/sigfile/AnnotationParser.java
@@ -158,7 +158,7 @@ public class AnnotationParser {
     }
 
     private void parseAnnData(AnnotationItem item, String specificData) {
-        item.setTarget(Integer.valueOf(specificData).intValue());
+        item.setTarget(Integer.valueOf(specificData));
     }
 
     private void parseAnnExData(AnnotationItemEx item, String specificData) {
